Moved the getRSS function from Browse.vue over to App.vue
